+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Add `sync_standby` as a valid replica type for `cluster_has_replica`. (contributed by @mattpoel)
 * Add a new service `cluster_has_scheduled_action` to warn of any scheduled switchover or restart.
 * Add options to `node_is_replica` to check specifically for a synchronous (`--is-sync`) or asynchronous node (`--is-async`).
+* Add `standby-leader` as a valid leader type for `cluster_has_leader`.
 
 ### Fixed